define and use comparePlacementSamples
