+++ frontend/src/components/placementVisuals/PlacementSamplesTable.tsx
@@ -9,7 +9,7 @@ import { fetchSamples } from "../../modules/cache/cache";
 import { selectCell, selectContainer } from "../../modules/placement/selectors";
 import { selectActiveDestinationContainer, selectActiveSourceContainer } from "../../modules/labworkSteps/selectors";
 import store from "../../store";
-import { compareArray, coordinatesToOffsets } from "../../utils/functions";
+import { comparePlacementSamples } from "../../utils/functions";
 export interface PlacementSamplesTableProps {
     container: string | null
 }
@@ -132,28 +132,7 @@ const PlacementSamplesTable = ({ container: containerName }: PlacementSamplesTab
     const sortedSamples: SortedSample[] = useMemo(() => {
         // const reverse = labworkSelectedSamples.reverse()
         const sortedSamples = [...samples]
-        sortedSamples.sort((a, b) => {
-            const MAX = 128
-
-            let orderA = MAX
-            let orderB = MAX
-
-            if (a.selected) orderA -= MAX/2
-            if (b.selected) orderB -= MAX/2
-
-            if (container && a.coordinates && b.coordinates) {
-                const aOffsets = coordinatesToOffsets(container.spec, a.coordinates)
-                const bOffsets = coordinatesToOffsets(container.spec, b.coordinates)
-                const arrayComparison = compareArray(aOffsets.reverse(), bOffsets.reverse())
-                if (arrayComparison > 0) orderB -= MAX/4
-                if (arrayComparison < 0) orderA -= MAX/4
-            }
-
-            if (a.name > b.name) orderB -= MAX/8
-            if (a.name < b.name) orderA -= MAX/8
-
-            return orderA - orderB
-        })
+        sortedSamples.sort((a, b) => comparePlacementSamples(a, b, container?.spec))
         return sortedSamples.reduce((sortedSamples, s) => {
             // s.sample also represents a sample from another container
             if (s.sample) {
